requirements.txt
rpi_lcd
smbus2